switched actions to doubles

+++ src/Arbiter.java
@@ -12,10 +12,10 @@ import orc.*;
 
 public class Arbiter extends java.lang.Thread {
 	public boolean running = true;
-	public float[] leftMotorAction = null;
+	public double[] leftMotorAction = null;
 	public float[] leftMotorWeight = null;
 	public float[] leftMotorLog = null;
-	public float[] rightMotorAction = null;
+	public double[] rightMotorAction = null;
 	public float[] rightMotorWeight = null;
 	public float[] rightMotorLog = null;
 	public double rollerAction = 0.0;
@@ -61,8 +61,8 @@ public class Arbiter extends java.lang.Thread {
 					maxval = leftMotorWeight[maxidx];
 				}
 			}
-			float lma = leftMotorWeight[maxidx];
-			float rma = rightMotorAction[maxidx];
+			double lma = leftMotorWeight[maxidx];
+			double rma = rightMotorAction[maxidx];
 			leftMotor.setPWM(bound(lma, 1.0f, -1.0f));
 			rightMotor.setPWM(bound(rma, 1.0f, -1.0f)*0.9f);
 			rollers.setPWM(rollerAction);
@@ -88,9 +88,9 @@ public class Arbiter extends java.lang.Thread {
 	public void setup(int numComrades) {
 		try {
 		o = new orc.Orc(java.net.Inet4Address.getByAddress(inet));
-		leftMotorAction = new float[numComrades];
+		leftMotorAction = new double[numComrades];
 		leftMotorWeight = new float[numComrades];
-		rightMotorAction = new float[numComrades];
+		rightMotorAction = new double[numComrades];
 		rightMotorWeight = new float[numComrades];
 		} catch (Exception e) {
 			e.printStackTrace();
